Use prefix iterators for Get/Has()
Pebble's Get implementation uses regular iterators under the hood, which
result in reading data out of sstables even when the key that we are looking
for does not exist (it will read the next key AND value).

This makes Has() and Get() calls very expensive for non existent keys
(equivalent to reading an existing one).

Per Pebble devs recommendation, we must use an iterator and SeekPrefixGE() to
make use of bloom filters, and bloom filters are built according to the Split
function in the Comparer, which is unset by default.

This PR forces setting the Split function in the options and uses
SeekPrefixGE() for Has()/Get(), which should make use of bloom filters.

The downside is that we do not allow custom Split functions, but we were not
taking advantage of them anyways. The second downside is that compactions will
call Split() for every item (to prime the bloom filters), so write performance
might be a bit worse.

Has() is still an expensive operation for existing keys, since the values are
read, but this cannot be really fixed.

For more details: https://github.com/cockroachdb/pebble/issues/2369