Do not allow batch to be reused after commit (#67)
By explicitly not allowing the Batch to be reused, the Batch resources can be recycled, after Commit is called, for use in a new Batch.

In general, a Batch should not be reuesd for this or any other datastore becuase it is not guaranteed to be safely reusable for all datastores. Therefore, making this Batch explicitly not reusable allows better resource management as well as returning a clear error when the Batch is mistakenly reused.